backend/requirements.txt
python-decouple
psycopg[binary]
pillow
django-cors-headers
djangorestframework
flake8
gunicorn